PP-884: avoid requesting user data when no account is set
