+++ src/init/getPeachInfo.ts
@@ -1,6 +1,5 @@
 import {
   setBuckets,
-  setClientServerTimeDifference,
   setDeprecatedBuckets,
   setLatestAppVersion,
   setMinAppVersion,
@@ -8,30 +7,11 @@ import {
   setPeachFee,
   setPeachPGPPublicKey,
 } from '../constants'
-import { tradeSummaryStore } from '../store/tradeSummaryStore'
 import { defaultAccount, updateTradingLimit } from '../utils/account'
-import { error, info } from '../utils/log'
-import { getContracts, getInfo, getOffers, getStatus, getTradingLimit } from '../utils/peachAPI'
+import { error } from '../utils/log'
+import { getInfo, getTradingLimit } from '../utils/peachAPI'
 import { sessionStorage } from '../utils/session'
-
-/**
- * Note: we estimate the time it took for the response to arrive from server to client
- * by dividing the round trip time in half
- * This is only an estimation as round trips are often asymmetric
- */
-const calculateClientServerTimeDifference = async () => {
-  const start = Date.now()
-  const [peachStatusResponse, peachStatusErr] = await getStatus({ timeout: 3000 })
-  const end = Date.now()
-  const roundTrip = (end - start) / 2
-
-  if (!peachStatusResponse || peachStatusErr) {
-    error('Error peach server info', JSON.stringify(peachStatusErr))
-    return
-  }
-
-  setClientServerTimeDifference(end - roundTrip - peachStatusResponse.serverTime)
-}
+import { calculateClientServerTimeDifference } from './calculateClientServerTimeDifference'
 
 /**
  * @description Method to fetch peach info and user trading limit and store values in constants
@@ -65,24 +45,4 @@ export const getPeachInfo = async (account?: Account): Promise<GetInfoResponse |
   }
 
   return peachInfo
-}
-
-/**
- * @description Method to fetch users offers and contracts
- */
-export const getTrades = async (): Promise<void> => {
-  const [offers, getOffersError] = await getOffers({})
-  if (offers) {
-    info(`Got ${offers.length} offers`)
-    tradeSummaryStore.getState().setOffers(offers)
-  } else if (getOffersError) {
-    error('Error', getOffersError)
-  }
-
-  const [contracts, err] = await getContracts({})
-  if (contracts) {
-    tradeSummaryStore.getState().setContracts(contracts)
-  } else if (err) {
-    error('Error', err)
-  }
 }
