+++ src/init/initApp.ts
@@ -1,9 +1,10 @@
 import { dataMigrationAfterLoadingAccount, dataMigrationBeforeLoadingAccount } from '../init/dataMigration'
 import events from '../init/events'
 import requestUserPermissions from '../init/requestUserPermissions'
-import { getPeachInfo, getTrades } from '../init/session'
 import userUpdate from '../init/userUpdate'
 import { account, loadAccount } from '../utils/account'
+import { getPeachInfo } from './getPeachInfo'
+import { getTrades } from './getTrades'
 
 /**
  * @description Method to initialize app by retrieving app session and user account
